+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 0.1.19
+
+### Added
+
+- PWA support to allow installing app to mobile
+
+### Fixed
+
+- disable today's earned SSR (#38)
+
 ## Version 0.1.18
 
 ### Added